First commit for the new Mac OS X project
Converted the C rendering functions to Objective-C / Cocoa.  Added in
the definitions for methods present in the super classes such as
drawRect:, mouseDown:, etc. The code to handle rotations, translations
and other matrix stuff isn’t in yet - that’s the stuff I’m working on
now. There’s a basic drawing function that clears the scene and puts a
triangle up there. It’s pretty rough right now, but things are drawing
on the screen.